Split package cli
