+++ packages/cli-proc/src/index.ts
@@ -1,0 +1,4 @@
+export { gcFont } from "./procs/gc";
+export { mergeFonts, MergeOption } from "./procs/merge";
+export { rebaseFont } from "./procs/rebase";
+export { subsetFont } from "./procs/subset";
