+++ packages/cli/src/cli-help/style.ts
@@ -5,6 +5,7 @@ const terminalSupportsUnicode =
     process.env.TERM_PROGRAM === "vscode" ||
     !!process.env.WT_SESSION; // Workaround https://github.com/microsoft/terminal/issues/1040
 
+// From strip-ansi
 const AnsiPattern = [
     "[\\u001B\\u009B][[\\]()#;?]*(?:(?:(?:[a-zA-Z\\d]*(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]*)*)?\\u0007)",
     "(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PR-TZcf-ntqry=><~]))"
