+++ packages/cli/src/index.ts
@@ -2,8 +2,6 @@ import { StartParseArgv } from "./argv-parser";
 import { createGrammar } from "./create-grammar";
 import { CliState } from "./state";
 
-export { CliProc } from "./create-grammar";
-
 export async function cliMain(argv: string[]) {
     const parse = StartParseArgv(argv, 2);
     const syntax = createGrammar();
