+++ packages/cli/src/syntax/actions/gc.ts
@@ -1,10 +1,8 @@
-import * as Ot from "@ot-builder/font";
-import * as Rectify from "@ot-builder/rectify-font";
+import { CliProc, Ot } from "ot-builder";
 import { ParseResult } from "../../argv-parser";
 import { CliHelpShower } from "../../cli-help";
 import { CliOptionStyle } from "../../cli-help/style";
 import { CliAction, Syntax } from "../../command";
-import { createSubsetRectifier } from "../../support/initial-visible-glyphs";
 
 export const GcSyntax: Syntax<null | CliAction> = {
     handle: st => {
@@ -16,7 +14,7 @@ export const GcSyntax: Syntax<null | CliAction> = {
             console.log(`Garbage collect ${entry}`);
 
             const gcBefore = entry.font.glyphs.decideOrder().length;
-            const gcResult = gcFont(entry.font, Ot.ListGlyphStoreFactory);
+            const gcResult = CliProc.gcFont(entry.font, Ot.ListGlyphStoreFactory);
             const gcAfter = gcResult.glyphs.decideOrder().length;
 
             state.push(entry.fill(gcResult));
@@ -28,14 +26,3 @@ export const GcSyntax: Syntax<null | CliAction> = {
         shower.indent("").message("Perform garbage collection of the font at the stack top.");
     }
 };
-
-export function gcFont<GS1 extends Ot.GlyphStore, GS2 extends Ot.GlyphStore>(
-    font: Ot.Font<GS1>,
-    gsf: Ot.GlyphStoreFactory<GS2>
-) {
-    const { glyphs, rectifier } = createSubsetRectifier(font, { has: () => true });
-
-    const font1 = { ...font, glyphs: gsf.createStoreFromList(glyphs) };
-    Rectify.rectifyFontGlyphReferences(rectifier, font1);
-    return font1;
-}
