+++ packages/cli/src/syntax/actions/rebase.ts
@@ -1,10 +1,8 @@
-import * as Ot from "@ot-builder/font";
-import * as Rectify from "@ot-builder/rectify-font";
+import { CliProc } from "ot-builder";
 import { ParseResult } from "../../argv-parser";
 import { CliHelpShower } from "../../cli-help";
 import { CliOptionStyle, CliParamStyle } from "../../cli-help/style";
 import { CliAction, Syntax } from "../../command";
-import { StdPointAttachRectifier } from "../../support/point-rectifier";
 
 export const RebaseSyntax: Syntax<null | CliAction> = {
     handle: st => {
@@ -16,7 +14,7 @@ export const RebaseSyntax: Syntax<null | CliAction> = {
             if (!entry) throw new RangeError("Stack size invalid. No font to rebase.");
             const newUpm = parseFloat(prArg.result) || entry.font.head.unitsPerEm;
             console.log(`Rebase ${entry} to ${newUpm}`);
-            rebaseFont(entry.font, newUpm);
+            CliProc.rebaseFont(entry.font, newUpm);
             state.push(entry);
         });
     },
@@ -25,28 +23,3 @@ export const RebaseSyntax: Syntax<null | CliAction> = {
         shower.indent("").message("Change the unit-per-em value of the font at the stack top.");
     }
 };
-
-export function rebaseFont<GS extends Ot.GlyphStore>(font: Ot.Font<GS>, newUpm: number) {
-    Rectify.rectifyFontCoords(
-        createAxisRectifier(),
-        createValueRectifier(newUpm, font.head.unitsPerEm),
-        new StdPointAttachRectifier(Rectify.PointAttachmentRectifyManner.TrustAttachment),
-        font
-    );
-    font.head.unitsPerEm = newUpm;
-}
-
-function createAxisRectifier(): Rectify.AxisRectifier {
-    return {
-        dim: a => a,
-        axis: a => a,
-        addedAxes: []
-    };
-}
-
-function createValueRectifier(newUpm: number, oldUpm: number): Rectify.CoordRectifier {
-    return {
-        coord: x => Ot.Var.Ops.scale(newUpm / oldUpm, x),
-        cv: x => Ot.Var.Ops.scale(newUpm / oldUpm, x)
-    };
-}
