+++ packages/cli/src/syntax/actions/subset.ts
@@ -1,10 +1,8 @@
-import * as Ot from "@ot-builder/font";
-import * as Rectify from "@ot-builder/rectify-font";
+import { CliProc, Ot } from "ot-builder";
 import { ParseResult } from "../../argv-parser";
 import { CliHelpShower } from "../../cli-help";
 import { CliOptionStyle, CliParamStyle } from "../../cli-help/style";
 import { CliAction, Syntax } from "../../command";
-import { createSubsetRectifier } from "../../support/initial-visible-glyphs";
 
 export const SubsetSyntax: Syntax<null | CliAction> = {
     handle: st => {
@@ -16,7 +14,11 @@ export const SubsetSyntax: Syntax<null | CliAction> = {
             if (!entry) throw new RangeError("Stack size invalid. No font to subset.");
             console.log(`Subset ${entry}`);
             const gcBefore = entry.font.glyphs.decideOrder().length;
-            const gcResult = subsetFont(entry.font, prArg.result, Ot.ListGlyphStoreFactory);
+            const gcResult = CliProc.subsetFont(
+                entry.font,
+                prArg.result,
+                Ot.ListGlyphStoreFactory
+            );
             const gcAfter = gcResult.glyphs.decideOrder().length;
 
             state.push(entry.fill(gcResult));
@@ -28,18 +30,3 @@ export const SubsetSyntax: Syntax<null | CliAction> = {
         shower.indent("").message("Subset the font at the stack top according to the text given.");
     }
 };
-
-export function subsetFont<GS1 extends Ot.GlyphStore, GS2 extends Ot.GlyphStore>(
-    font: Ot.Font<GS1>,
-    text: string,
-    gsf: Ot.GlyphStoreFactory<GS2>
-) {
-    const { glyphs, rectifier } = createSubsetRectifier(
-        font,
-        new Set([...text].map(s => s.codePointAt(0)!))
-    );
-
-    const font1 = { ...font, glyphs: gsf.createStoreFromList(glyphs) };
-    Rectify.rectifyFontGlyphReferences(rectifier, font1);
-    return font1;
-}
