+++ packages/cli/src/syntax/document/help.ts
@@ -1,10 +1,7 @@
-import * as Ot from "@ot-builder/font";
-import * as Rectify from "@ot-builder/rectify-font";
 import { ParseResult } from "../../argv-parser";
 import { CliHelpShower } from "../../cli-help";
 import { CliCmdStyle, CliOptionStyle } from "../../cli-help/style";
 import { CliAction, Syntax } from "../../command";
-import { createSubsetRectifier } from "../../support/initial-visible-glyphs";
 
 export const HelpSyntax: Syntax<null | CliAction[]> = {
     handle: (st, sy) => {
@@ -26,14 +23,3 @@ export const HelpSyntax: Syntax<null | CliAction[]> = {
         shower.indent("").message("Display help message.");
     }
 };
-
-export function gcFont<GS1 extends Ot.GlyphStore, GS2 extends Ot.GlyphStore>(
-    font: Ot.Font<GS1>,
-    gsf: Ot.GlyphStoreFactory<GS2>
-) {
-    const { glyphs, rectifier } = createSubsetRectifier(font, { has: () => true });
-
-    const font1 = { ...font, glyphs: gsf.createStoreFromList(glyphs) };
-    Rectify.rectifyFontGlyphReferences(rectifier, font1);
-    return font1;
-}
