+++ packages/demo/tsconfig.json
@@ -97,7 +97,7 @@
       "path": "../rectify-font/tsconfig.json"
     },
     {
-      "path": "../cli/tsconfig.json"
+      "path": "../cli-proc/tsconfig.json"
     },
     {
       "path": "../index/tsconfig.json"
