+++ packages/demo/tsconfig.prod.json
@@ -97,7 +97,7 @@
       "path": "../rectify-font/tsconfig.prod.json"
     },
     {
-      "path": "../cli/tsconfig.prod.json"
+      "path": "../cli-proc/tsconfig.prod.json"
     },
     {
       "path": "../index/tsconfig.prod.json"
